feat: add pull request preview with graphql (#42)
On this PR;
- Create a docker-compose inside `.preview` that builds and runs a
environment.
- Add `pullpreview` CI action to deploy PR previews to AWS Lightsail. 
- Add rewrites inside the next configuration to enable the project to
contain a single entry point. This is going to be used only on the
preview envs.
- Add a start action on turbo that runs the start on the `graphql` and
`app` packages.

### About PullPreview service

The PR preview uses Lightsail from AWS in combination with
https://github.com/pullpreview/action does all the work of creating the
instance, pulling the code there, and running the
`.preview/docker-compose.yaml` They also assign a pretty URL using their
domain using [sslip](https://github.com/pullpreview/docker-sslip).

This `pullpreview` solution is going to be tested for the next couple of
days to evaluate if we are going to keep it or not. Leveraging the 30
days trial, they offer on the license.